requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
selenium==4.21.0
webdriver-manager==4.0.1
pydantic==2.10.6  # Updated to be compatible with pydantic-settings
pdfplumber==0.11.0
python-magic==0.4.27
pydantic-settings==2.8.1
mangum==0.17.0
reportlab==4.1.0  # For generating PDF files